Feb 26 Updates
Added a home page and changed the general structure of the app flow. Instead of a SPA, I decided it might be better to have a homepage with a main section and some basic descriptions of the app's use and purpose. Later the home page will also serve as the place where edits to pane pricing happens.